+++ src/App.js
@@ -1,26 +1,12 @@
 import React from 'react';
 import { withStyles } from 'material-ui/styles';
-import Drawer from 'material-ui/Drawer';
-import AppBar from 'material-ui/AppBar';
-import Toolbar from 'material-ui/Toolbar';
-import Typography from 'material-ui/Typography';
-import Divider from 'material-ui/Divider';
-import Input from 'material-ui/Input';
 
 import { chats, messages } from './mock-data';
 
-import List, { ListItem, ListItemText } from 'material-ui/List';
-import Avatar from 'material-ui/Avatar';
-import Button from 'material-ui/Button';
-import RestoreIcon from 'material-ui-icons/Restore';
-import ExploreIcon from 'material-ui-icons/Explore';
-import AddIcon from 'material-ui-icons/Add';
-import BottomNavigation, { BottomNavigationAction } from 'material-ui/BottomNavigation';
-import Paper from 'material-ui/Paper';
-//---------------------------------
-import classnames from 'classnames'
-import titleInitials from './utils/title-initials'
 
+import Sidebar from './components/sidebar'
+import ChatHeader from './components/ChatHeader'
+import Chat from './components/Chat'
 const styles = theme => ({
   root: {
     position: 'relative',
@@ -33,26 +19,7 @@ const styles = theme => ({
     position: 'fixed',
     width: `calc(100% - 320px)`,
   },
-  drawerPaper: {
-    position: 'relative',
-    height: '100%',
-    width: 320,
-  },
-  drawerHeader: {
-    ...theme.mixins.toolbar,
-    paddingLeft: theme.spacing.unit * 3,
-    paddingRight: theme.spacing.unit * 3,
-  },
-  chatsList: {
-    height: 'calc(100% - 56px)',
-    overflowY: 'scroll',
-  },
-  newChatButton: {
-    position: 'absolute',
-    left: 'auto',
-    right: theme.spacing.unit * 3,
-    bottom: theme.spacing.unit * 3 + 48, // + bottom navigation
-  },
+  
   chatLayout: {
     display: 'flex',
     justifyContent: 'center',
@@ -60,44 +27,7 @@ const styles = theme => ({
     paddingTop: '64px',
     height: '100%',
     overflow: 'hidden',
-  },
-  messagesWrapper: {
-    overflowX: 'scroll',
-    height: '100%',
-    width: '100%',
-    paddingTop: theme.spacing.unit * 3,
-    paddingBottom: '120px',
-  },
-  messageInputWrapper: {
-    position: 'fixed',
-    left: 'auto',
-    right: 0,
-    bottom: 0,
-    width: `calc(100% - 400px)`,
-    padding: theme.spacing.unit * 3,
-  },
-  messageInput: {
-    padding: theme.spacing.unit * 2,
-  },
-  messageWrapper: {
-    display: 'flex',
-    justifyContent: 'flex-start',
-    alignItems: 'center',
-    padding: `${theme.spacing.unit}px ${theme.spacing.unit * 3}px`,
-  },
-  messageWrappperFromMe: {
-    justifyContent: 'flex-end',
-  },
-  message: {
-    maxWidth: '70%',
-    minWidth: '10%',
-    padding: theme.spacing.unit,
-    marginLeft: theme.spacing.unit * 2,
-  },
-  messageFromMe: {
-    marginRight: theme.spacing.unit * 2,
-    backgroundColor: '#e6dcff'
-  },
+  }
 });
 class App extends React.Component {
   constructor(props){
@@ -113,7 +43,7 @@ class App extends React.Component {
   }
 
   scrollDownHistory() {
-    debugger;
+  
     const messagesWrapper = this.refMessagesWrapper.current;// this.refs.messagesWrapper;
     if (messagesWrapper) {
       messagesWrapper.scrollTop =messagesWrapper.scrollHeight;
@@ -128,95 +58,11 @@ class App extends React.Component {
     return (
 
       <div className={classes.root}>
-        <AppBar
-          position="absolute"
-          className={classes.appBar}
-        >
-          <Toolbar>
-            <Typography variant="title" color="inherit" noWrap>
-              DogeCodes React Chat
-              </Typography>
-          </Toolbar>
-        </AppBar>
-        <Drawer
-          variant="permanent"
-          classes={{
-            paper: classes.drawerPaper,
-          }}
-        >
-          {/*   <div className={classes.toolbar} /> */}
-          <div className={classes.drawerHeader}>
-            <Input
-              placeholder="Search chats..."
-              className={classes.input}
-              inputProps={{
-                'aria-label': 'Description',
-              }}
-            /></div>
-          <Divider />
-          <List className={classes.chatsList}>
-            {chats.map((chat, index) => (
-              <ListItem key={index} button>
-                <Avatar>{titleInitials(chat.title)}</Avatar>
-                <ListItemText primary={chat.title} />
-              </ListItem>
-            ))}
 
-          </List>
-          <Button
-            variant="fab"
-            color="primary"
-            className={classes.newChatButton}
-          >
-            <AddIcon />
-          </Button>
+<ChatHeader/>
+        <Sidebar chats={chats}/>
+        <Chat messages={messages}/>
 
-
-          <BottomNavigation showLabels>
-            <BottomNavigationAction label="My Chats" icon={<RestoreIcon />} />
-            <BottomNavigationAction label="Explore" icon={<ExploreIcon />} />
-          </BottomNavigation>
-         
-        </Drawer>
-
-        <main className={classes.chatLayout}  >
-          <div className={classes.messagesWrapper} ref={this.refMessagesWrapper} >
-            {messages && messages.map((message, index) => {
-              const isMessageFromMe = message.sender === 'me';
-
-              const userAvatar = (
-                <Avatar>
-                  {titleInitials(message.sender)}
-                </Avatar>
-              );
-
-              return (
-                // <div className={classNames(classes.messageWrapper, isMessageFromMe && classes.messageWrappperFromMe)}>
-                <div key={index} className={classnames(
-                  classes.messageWrapper,
-                  isMessageFromMe && classes.messageWrappperFromMe)}>
-                  {!isMessageFromMe && userAvatar}
-                  <Paper className={classnames(
-                    classes.message,
-                    isMessageFromMe && classes.messageFromMe)}>
-                    <Typography variant="caption">
-                      {message.sender}
-                    </Typography>
-                    <Typography variant="body1">
-                      {message.content}
-                    </Typography>
-                  </Paper>
-                  {isMessageFromMe && userAvatar}
-                </div>
-              );
-            })}
-          </div>
-          <div className={classes.messageInputWrapper}>
-            <Paper className={classes.messageInput} elevation={6}>
-              <Input fullWidth placeholder="Type your message…"/>
-            </Paper>
-          </div>
-        </main>
       </div>
 
     );
