+++ src/components/ChatHeader.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import { withStyles } from 'material-ui/styles';
+import AppBar from 'material-ui/AppBar';
+import Toolbar from 'material-ui/Toolbar';
+import Typography from 'material-ui/Typography';
+
+const styles = theme => ({
+  appBar: {
+    position: 'fixed',
+    width: `calc(100% - 320px)`,
+  },
+});
+
+const ChatHeader = ({ classes }) => (
+  <AppBar color="primary" className={classes.appBar}>
+    <Toolbar>
+      <Typography variant="title" color="inherit" noWrap>
+        DogeCodes React Chat
+      </Typography>
+    </Toolbar>
+  </AppBar>
+);
+
+export default withStyles(styles)(ChatHeader);
