+++ src/components/sidebar.js
@@ -1,0 +1,92 @@
+import React from 'react';
+import { withStyles } from 'material-ui/styles';
+
+import Drawer from 'material-ui/Drawer';
+import Divider from 'material-ui/Divider';
+import TextField from 'material-ui/TextField';
+import List, {ListItem,ListItemText} from 'material-ui/List';
+import BottomNavigation, { BottomNavigationAction } from 'material-ui/BottomNavigation';
+
+import RestoreIcon from 'material-ui-icons/Restore';
+import ExploreIcon from 'material-ui-icons/Explore';
+import titleInitials from  '../utils/title-initials';
+
+import Avatar from 'material-ui/Avatar';
+import Button from 'material-ui/Button';
+
+
+import AddIcon from 'material-ui-icons/Add';
+
+
+const styles = theme => ({
+  drawerPaper: {
+    position: 'relative',
+    height: '100%',
+    width: 320,
+  },
+  drawerHeader: {
+    ...theme.mixins.toolbar,
+    paddingLeft: theme.spacing.unit * 3,
+    paddingRight: theme.spacing.unit * 3,
+  },
+  chatsList: {
+    height: 'calc(100% - 56px)',
+    overflowY: 'scroll',
+
+  },
+  newChatButton: {
+    position: 'absolute',
+    left: 'auto',
+    right: theme.spacing.unit * 3,
+    bottom: theme.spacing.unit * 3 + 48, // + bottom navigation
+  },
+});
+
+
+const Sidebar = ({ classes, chats }) => (
+  <Drawer
+  variant="permanent"
+  classes={{
+    paper: classes.drawerPaper,
+  }}
+>
+  {/*   <div className={classes.toolbar} /> */}
+  <div className={classes.drawerHeader}>
+      <TextField
+        fullWidth
+        margin="normal"
+        placeholder="Search chats..."
+      />
+    </div>
+  <Divider />
+  <List className={classes.chatsList}>
+    {chats.map((chat, index) => (
+      <ListItem key={index} button>
+        <Avatar>{titleInitials(chat.title)}</Avatar>
+        <ListItemText primary={chat.title} />
+      </ListItem>
+    ))}
+
+  </List>
+  <Button
+    variant="fab"
+    color="primary"
+    className={classes.newChatButton}
+  >
+    <AddIcon />
+  </Button>
+
+
+  <BottomNavigation showLabels>
+    <BottomNavigationAction label="My Chats" icon={<RestoreIcon />} />
+    <BottomNavigationAction label="Explore" icon={<ExploreIcon />} />
+  </BottomNavigation>
+ 
+</Drawer>
+);
+
+
+
+
+
+export default withStyles(styles)(Sidebar);
